requirements.txt
Django==1.7.8
django-multiselectfield==0.1.3
gevent==1.0.2
greenlet==0.4.7
logutils==0.3.3
msgpack-python==0.4.6
MySQL-python==1.2.5
Pillow==2.8.2
python-dateutil==2.4.2
python-prctl==1.6.1
pyzmq==14.7.0
six==1.9.0
urllib3==1.10.4
wheel==0.24.0
zerorpc==0.5.1
argparse==1.2.1
wsgiref==0.1.2
